Rename the generators module to charges and rebase the BCC model into it.
