--- openff/recharge/models/__init__.py
@@ -1,3 +1,0 @@
-from openff.recharge.models.models import BondChargeCorrection
-
-__all__ = [BondChargeCorrection]
