+++ openff/recharge/tests/charges/test_bcc.py
@@ -2,12 +2,9 @@ import numpy
 import pytest
 from openeye import oechem, oequacpac
 
+from openff.recharge.charges.bcc import AM1BCC
+from openff.recharge.charges.exceptions import OEQuacpacError, UnableToAssignChargeError
 from openff.recharge.conformers.conformers import OmegaELF10
-from openff.recharge.generators.bcc import AM1BCC
-from openff.recharge.generators.exceptions import (
-    OEQuacpacError,
-    UnableToAssignChargeError,
-)
 from openff.recharge.utilities.exceptions import MissingConformersError
 from openff.recharge.utilities.openeye import call_openeye, smiles_to_molecule
